Conflicts:
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Afrikaans/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Belarusian/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Bulgarian/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Catalan/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Chinese (Simple)/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Czech/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Danish/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Dutch/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/English/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Finnish/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/French/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/German/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Greek/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Hebrew/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Hungarian/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Italian/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Korean/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Lithuanian/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Macedonian/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Polish/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Portuguese (Brazil)/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Portuguese/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Romanian/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Russian/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Slovak/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Slovenian/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Spanish (Argentina)/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Spanish (Mexico)/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Spanish/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Swedish/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/language/Thai/strings.xml
	addons/metadata.musicvideos.theaudiodb.com/resources/settings.xml
	configure.in
	lib/libdvd/libdvdcss/configure.ac
	media/Splash.png
	project/Win32BuildSetup/buildpvraddons.bat
	system/peripherals.xml
	tools/android/depends/Makefile.in
	tools/android/depends/Makefile.include.in
	tools/android/depends/xbmc-pvr-addons/Makefile
	tools/android/packaging/xbmc/AndroidManifest.xml
	tools/darwin/depends/xbmc-pvr-addons/Makefile
	tools/darwin/packaging/xbmc-atv2/mkdeb-xbmc-atv2.sh
	tools/darwin/packaging/xbmc-ios/mkdeb-xbmc-ios.sh
	xbmc/Application.cpp
	xbmc/ApplicationMessenger.cpp
	xbmc/GUIInfoManager.h
	xbmc/cores/AudioEngine/Engines/CoreAudio/CoreAudioDevice.cpp
	xbmc/cores/AudioEngine/Engines/CoreAudio/CoreAudioHardware.cpp
	xbmc/cores/AudioEngine/Engines/SoftAE/SoftAE.cpp
	xbmc/cores/IPlayer.h
	xbmc/cores/VideoRenderers/BaseRenderer.cpp
	xbmc/cores/amlplayer/AMLPlayer.cpp
	xbmc/cores/amlplayer/AMLPlayer.h
	xbmc/cores/dvdplayer/DVDCodecs/Video/DVDVideoCodecVDA.cpp
	xbmc/cores/dvdplayer/DVDPlayer.cpp
	xbmc/cores/dvdplayer/DVDPlayer.h
	xbmc/cores/omxplayer/OMXAudio.cpp
	xbmc/cores/omxplayer/OMXPlayer.h
	xbmc/cores/omxplayer/OMXPlayerVideo.cpp
	xbmc/cores/omxplayer/OMXVideo.cpp
	xbmc/cores/omxplayer/OMXVideo.h
	xbmc/guilib/GUITextLayout.cpp
	xbmc/guilib/JpegIO.cpp
	xbmc/interfaces/json-rpc/PlayerOperations.cpp
	xbmc/interfaces/legacy/File.cpp
	xbmc/music/MusicInfoLoader.cpp
	xbmc/music/infoscanner/MusicInfoScanner.cpp
	xbmc/osx/HotKeyController.m
	xbmc/osx/Info.plist
	xbmc/peripherals/devices/PeripheralCecAdapter.cpp
	xbmc/peripherals/devices/PeripheralCecAdapter.h
	xbmc/peripherals/devices/PeripheralNyxboard.cpp
	xbmc/powermanagement/PowerManager.cpp
	xbmc/pvr/PVRManager.h
	xbmc/utils/JobManager.cpp
	xbmc/win32/XBMC_PC.rc
	xbmc/windowing/egl/EGLNativeTypeRaspberryPI.cpp
	xbmc/windowing/egl/EGLWrapper.cpp
	xbmc/windowing/egl/WinSystemEGL.cpp
	xbmc/windowing/osx/WinSystemOSX.mm

+++ xbmc/interfaces/json-rpc/ApplicationOperations.cpp
@@ -1,5 +1,5 @@
 /*
- *      Copyright (C) 2005-2012 Team XBMC
+ *      Copyright (C) 2005-2013 Team XBMC
  *      http://www.xbmc.org
  *
  *  This Program is free software; you can redistribute it and/or modify
@@ -54,7 +54,7 @@ JSONRPC_STATUS CApplicationOperations::SetVolume(const CStdString &method, ITran
   bool up = false;
   if (parameterObject["volume"].isInteger())
   {
-    int oldVolume = g_application.GetVolume();
+    int oldVolume = (int)g_application.GetVolume();
     int volume = (int)parameterObject["volume"].asInteger();
   
     g_application.SetVolume((float)volume, true);
