+++ xbmc/video/FFmpegVideoDecoder.cpp
@@ -1,5 +1,5 @@
 /*
- *      Copyright (C) 2005-2012 Team XBMC
+ *      Copyright (C) 2005-2013 Team XBMC
  *      http://www.xbmc.org
  *
  *  This Program is free software; you can redistribute it and/or modify
@@ -295,7 +295,7 @@ bool FFmpegVideoDecoder::nextFrame( CBaseTexture * texture )
       // Did we get a video frame?
       if ( frameFinished )
       {
-        if ( packet.dts != AV_NOPTS_VALUE )
+        if ( packet.dts != (int64_t)AV_NOPTS_VALUE )
 	  m_lastFrameTime = packet.dts * av_q2d( m_pFormatCtx->streams[ m_videoStream ]->time_base );
         else
 	   m_lastFrameTime = 0.0;
